Get rid of superfluous file
